refactor: src/components directory refactoring
src/components 디렉터리 구조 및 이름 변경
디렉터리 구조 변경으로 인해 영향받은 코드들 일괄 수정.
