+++ src/components/CompButtonLight/CompButtonLight.scss
