+++ src/components/CompTestNav/CompTestNav.scss
@@ -1,4 +1,4 @@
-.Nav {
+.CompTestNav {
   position: fixed;
   right: 30px;
   bottom: 30px;
