+++ src/components/CompTestNav/index.jsx
@@ -1,0 +1,3 @@
+import CompTestNav from './CompTestNav';
+
+export default CompTestNav;
