--- src/components/test/Nav/index.jsx
@@ -1,3 +1,0 @@
-import Nav from './Nav';
-
-export default Nav;
